### 2017-11-30 #477 @whedon
[PRE REVIEW]: CoreRobotics: An object-oriented C++ library with cross-language wrappers for cross-platform robot control
**Submitting author:** @CoreRobotics (<a href="http://orcid.org/0000-0003-1160-0426">Parker Owan</a>)
**Repository:** <a href="https://github.com/CoreRobotics/CoreRobotics" target ="_blank">CoreRobotics/CoreRobotics</a>
**Version:** v0.9.1
