### 2017-11-30 #477 @CoreRobotics
Hello,

I believe the repository link of the original submission is broken.  The repository is maintained at gitlab, however a mirror is on Github.  The repository for this review needs to be updated to this link: [https://github.com/CoreRobotics/CoreRobotics](https://github.com/CoreRobotics/CoreRobotics)

From the list, @mosteo seems to be an appropriate reviewer.  @jordigh as well.
