### 2017-12-01 #477 @arfon
> I believe the repository link of the original submission is broken. The repository is maintained at gitlab, however a mirror is on Github. The repository for this review needs to be updated to this link: https://github.com/CoreRobotics/CoreRobotics

Thanks for spotting this. It's now fixed in this review issue.

> From the list, @mosteo seems to be an appropriate reviewer. @jordigh as well.

Thanks for the suggestions. We'll need an editor to assign themselves first before we can get going with the reviewers.
