### 2017-12-01 1798ea0
Updated documentation
