### 2017-12-01 #477 @danielskatz
@bmagyar - would you be willing to be on the other side of a JOSS paper and do a review?

@mosteo and @jordigh - would one of you be willing to review this as well?

See http://joss.theoj.org/about#reviewer_guidelines for JOSS reviewer guidelines.
