### 2017-12-05 #477 @danielskatz
@CoreRobotics - Do you have any other reviewer suggestions?  Such as someone else in the robotics community who knows software and is not conflicted with your or your co-authors, but perhaps isn't yet involved with JOSS?  I would like to find a second reviewer.
