### 2017-12-07 #477 @amjaeger17
Hi,
Yeah, I am interested, and can definetly help check off some of the boxes.
