### 2017-12-07 #489 @whedon
[REVIEW]: CoreRobotics: An object-oriented C++ library with cross-language wrappers for cross-platform robot control
**Submitting author:** @CoreRobotics (<a href="http://orcid.org/0000-0003-1160-0426">Parker Owan</a>)
**Repository:** <a href="https://github.com/CoreRobotics/CoreRobotics" target ="_blank">https://github.com/CoreRobotics/CoreRobotics</a>
**Version:** v0.9.1
**Reviewer:** @bmagyar @amjaeger17
