### 2017-12-07 #489 @danielskatz
@bmagyar & @amjaeger17 - this is ready to go

@amjaeger17 - you will have to accept the invitation to join the JOSS reviewers community: https://github.com/openjournals/joss-reviews/invitations before you will be able to check off items.  Once you do so, I will also assign you to the GitHub issue.

Please use the two checklists above (one for each of you) to carry out your reviews.

An informal guideline is that we would like your review in 2 weeks, but sooner (or later) are also ok.

The reviewer guidelines are available here: http://joss.theoj.org/about#reviewer_guidelines

Please avoid lengthy details of difficulties in this review thread. Instead, please create a new issue in the target repository and link to those issues (especially acceptance-blockers) in this review thread. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

Any questions/concerns, please let me know.

Thanks!!
