### 2017-12-07 c254f89
Fixed a comma issue in the paper.bib and added JOSS badge to Readme
