+++ paper.bib
@@ -1,8 +1,8 @@
 @misc{ros,
-    author = {{The Open Source Robotics Foundation}},
+    author = {The Open Source Robotics Foundation},
     title = {Robot Operating System (ROS)},
     year = {2017},
-    howpublished = {\url{http://www.ros.org/}}
+    url = {http://www.ros.org/}
 }
 
 @book{kreyszig,
@@ -10,13 +10,13 @@
     author = {E. Kreyszig},
     edition = {10th},
     publisher = {John Wiley {\&} Sons},
-    title = {{Advanced Engineering Mathematics}},
+    title = {Advanced Engineering Mathematics},
     year = {2011}
 }
 
 @book{craig,
     author = {J. Craig},
-    title = {{Introduction to Robotics: Mechanics and Control}},
+    title = {Introduction to Robotics: Mechanics and Control},
     edition = {3rd},
     publisher = {Pearson},
     year = {2004}
@@ -26,40 +26,40 @@
     author = {R. Murray, Z. Li, and S. Sastry},
     edition = {1},
     publisher = {CRC Press},
-    title = {{A Mathematical Introduction to Robot Manipulation}},
-    year = {1993}
-    howpublished = {\url{http://www.cds.caltech.edu/~murray/mlswiki/index.php/Main_Page}}
+    title = {A Mathematical Introduction to Robot Manipulation},
+    year = {1993},
+    url = {http://www.cds.caltech.edu/~murray/mlswiki/index.php/Main_Page}
 }
 
 @article{DLS,
     author = {A. Aristidou and J. Lasenby},
-    title = {{Inverse Kinematics: a review of existing techniques and introduction of a new fast iterative solver}},
+    title = {Inverse Kinematics: a review of existing techniques and introduction of a new fast iterative solver},
     publisher = {University of Cambridge, Technical Report},
     year = {2009},
-    howpublished = {\url{http://publications.eng.cam.ac.uk/324095/}}
+    url = {http://publications.eng.cam.ac.uk/324095/}
 }
 
 @article{NS,
     author = {O. Khatib},
-    title = {{Advanced Robotic Manipulation}},
+    title = {Advanced Robotic Manipulation},
     publisher = {Stanford Lecture Notes (CS327A)},
     year = {2005},
-    howpublished = {\url{http://www.in.tum.de/fileadmin/user_upload/Lehrstuehle/Lehrstuhl_XXIII/AdvancedRoboticManipulation.pdf}}
+    url = {http://www.in.tum.de/fileadmin/user_upload/Lehrstuehle/Lehrstuhl_XXIII/AdvancedRoboticManipulation.pdf}
 }
 
 @article{hessian,
     author = {A. Hourtash},
-    title = {{The Kinematic Hessian and Higher Derivatives}},
+    title = {The Kinematic Hessian and Higher Derivatives},
     publisher = {IEEE International Symposium on Computational Intelligence in Robotics and
 Automation},
     year = {2005},
     pages = {169--174},
-    howpublished = {\url{http://doi.org/10.1109/CIRA.2005.1554272}}
+    url = {http://doi.org/10.1109/CIRA.2005.1554272}
 }
 
 @article{minJerk,
     author = {N. Hogan},
-    title = {{Adaptive control of mechanical impedance by coactivation of antagonist muscles}},
+    title = {Adaptive control of mechanical impedance by coactivation of antagonist muscles},
     publisher = {IEEE Trans. on Automatic Control AC-29},
     year = {1984},
     pages = {681--690}
@@ -71,7 +71,7 @@ Automation},
     isbn = {978-0-262-20162-9},
[…]
