### 2017-12-07 #489 @CoreRobotics
Not sure what's going on with the pdf generator.  We have a pdf compiled from latex in the repository root, same content as the paper.md
