### 2017-12-07 #489 @arfon
> Not sure what's going on with the pdf generator. We have a pdf compiled from latex in the repository root, same content as the paper.md

I found a weird edge-case in the code that I just fixed. Looks like the paper is compiling OK now.
