### 2017-12-12 CoreRobotics/CoreRobotics#1 @bmagyar
[joss-review] An example application & figure in the paper would help quickly assessing the core functionalities of the library
...ideally with a real robot.
