### 2017-12-12 CoreRobotics/CoreRobotics#2 @bmagyar
[joss-review] Where are the tests?
I've found a few files named `test_X` in the `utils` folder but they are not really tests, they just print a bunch of stuff. I ran a `grep` for asserts but could only find one.

Please use a C++ testing framework to implement functional tests with asserts. GTest is usually a good choice.
