### 2017-12-12 CoreRobotics/CoreRobotics#3 @bmagyar
[joss-review] Shipping custom version of boost and Eigen?
While this seems to make life simpler at first, these dependencies should be installed separately.

I am unfamiliar with how deployment goes on Windows systems but both Linux and OSX has package managers where these dependencies can be grabbed from. Dragging around copies of someone else's software is not too elegant. The current solution locks one in with the specific version of boost which if used in conjunction with other libraries (a ROS wrapper around a controller for instance)  collisions would be very likely to occur.

My proposed solution is to move the external dependencies to a different repository which would only be downloaded if the person is on Windows, Linux and OSX systems can use the system default (apt or brew installated) versions of these libraries.
