### 2017-12-12 CoreRobotics/CoreRobotics#4 @bmagyar
[joss-review] Please don't define PI again
https://github.com/CoreRobotics/CoreRobotics/blob/master/src/math/CRMath.hpp#L98

https://stackoverflow.com/questions/6563810/m-pi-works-with-math-h-but-not-with-cmath-in-visual-studio
