### 2017-12-12 CoreRobotics/CoreRobotics#5 @bmagyar
[joss-review] Math library could be split into more files
Currently it's all dumped into a single file while I can see at least 3 separate categories:
* angle-related things
* integration methods
* matrix operations
