### 2017-12-12 CoreRobotics/CoreRobotics#6 @bmagyar
[joss-review] Avoid passing around CRManipulator raw pointers
Use any smart pointer from C++11 or (better suited for this use-case) const references to such an object.
