### 2017-12-12 CoreRobotics/CoreRobotics#7 @bmagyar
[joss-review] Simplification for code style
...although magnitude of this change may be big.

Example:
`CoreRobotics::CRClock`

Why the CR prefix? The namespace is already an excellent tag for it. Probably the namespaces were added after the core classes of the library but this simplification step wasn't done. This affect all classes I believe.
