### 2017-12-12 #489 @amjaeger17
@danielskatz for  documentation is there an expectation that there is a static/pregenerated api documentation? Or is a Doxygen config file acceptable? - Unclear from JOSS guidlines although most repos seem to have a link to docs.
