### 2017-12-12 #489 @danielskatz
Just to take one checkbox as an example:
 >Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I would expect that if you check this box, there is either hand-written or auto-generated text in the repo or after the software is built that includes such examples.

Does that answer your question?
