### 2017-12-18 #489 @bmagyar
I remember ticking some boxes, perhaps they didn't get saved or I was a bit of a noob and ticked the boxes belonging to @amjaeger17 . I've actually opened a [bunch of issues](https://github.com/CoreRobotics/CoreRobotics/issues) already
