### 2017-12-18 #489 @danielskatz
Thanks - as you open issues, please put a note in this issue saying that you have done so, and ideally linking to those issues as appropriate, so a future reader (or the current editor) can look at this issue and understand the process of the overall review.
