### 2017-12-18 #489 @bmagyar
Good point. Here they come:

* https://github.com/CoreRobotics/CoreRobotics/issues/1
* https://github.com/CoreRobotics/CoreRobotics/issues/2
* https://github.com/CoreRobotics/CoreRobotics/issues/3
* https://github.com/CoreRobotics/CoreRobotics/issues/4
* https://github.com/CoreRobotics/CoreRobotics/issues/5
* https://github.com/CoreRobotics/CoreRobotics/issues/6
* https://github.com/CoreRobotics/CoreRobotics/issues/7
