### 2017-12-19 CoreRobotics/CoreRobotics#7 @CoreRobotics
Inclusion of CR was to avoid name conflicts before the namespace was introduced.  Unfortunately, this change would significantly impact existing projects.
