### 2017-12-19 058af7b
Updated to remove CR_PI constant, using M_PI instead
