+++ utils/test_CRNullSpace.cpp
@@ -104,10 +104,10 @@ void test_CRNullSpace(void) {
 	// Set Euler Angles and Position Offsets
 	//							  dx		dy		dz		tx				ty		tz
 	F0->setPositionAndOrientation(0.0,		0.0,	d1,		0.0,			0.0,	0.0);
-	F1->setPositionAndOrientation(0.0,		0.0,	0.0,	CR_PI/2.0,		0.0,	0.0);
+    F1->setPositionAndOrientation(0.0,		0.0,	0.0,	M_PI/2.0,		0.0,	0.0);
 	F2->setPositionAndOrientation(d3,		d4,		d2,		0.0,			0.0,	0.0);
 	F3->setPositionAndOrientation(d5,		0.0,	-d6,	0.0,			0.0,	0.0);
-	F4->setPositionAndOrientation(d7,		0.0,	-d8,	-CR_PI/2.0,		0.0,	0.0);
+	F4->setPositionAndOrientation(d7,		0.0,	-d8,	-M_PI/2.0,		0.0,	0.0);
 	F5->setPositionAndOrientation(d9,		d10,	0.0,	0.0,			0.0,	0.0);
 	F6->setPositionAndOrientation(0.0,		0.0,	0.0,	0.0,			0.0,	0.0);
 
@@ -133,7 +133,7 @@ void test_CRNullSpace(void) {
 	CRFrameEuler* Tool = new CRFrameEuler();
 	Tool->setFreeVariable(CR_EULER_FREE_NONE);
 	Tool->setMode(CR_EULER_MODE_XYZ);
-	Tool->setPositionAndOrientation(90.21, 0.0, -107.43, -CR_PI / 8.0, 0.0, 0.0);
+	Tool->setPositionAndOrientation(90.21, 0.0, -107.43, -M_PI / 8.0, 0.0, 0.0);
 	int toolIndex = MyRobot->addTool(linkIndex6, Tool);
 
 	// Initialize the solver
@@ -141,7 +141,7 @@ void test_CRNullSpace(void) {
 
 	// Set the robot orientation
 	Eigen::VectorXd InitJoints(6);
-	InitJoints << 0.0, 2.0, 0.0, -2, 0.0, CR_PI / 8.0;
+	InitJoints << 0.0, 2.0, 0.0, -2, 0.0, M_PI / 8.0;
 	MyRobot->setConfiguration(InitJoints);
 
 	// Compute the initial tool pose
