+++ src/physics/CRFrame.cpp
@@ -242,7 +242,7 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI;
+                b = M_PI;
                 a = -atan2(-r01, r00);
                 g = 0;
             }
@@ -265,7 +265,7 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI;
+                b = M_PI;
                 a = -atan2(-r12, r11);
                 g = 0;
             }
@@ -288,7 +288,7 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI;
+                b = M_PI;
                 a = -atan2(-r20, r22);
                 g = 0;
             }
@@ -311,7 +311,7 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI;
+                b = M_PI;
                 a = -atan2(r10, r11);
                 g = 0;
             }
@@ -334,7 +334,7 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI;
+                b = M_PI;
                 a = -atan2(r21, r22);
                 g = 0;
             }
@@ -357,7 +357,7 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI;
+                b = M_PI;
                 a = -atan2(r02, r00);
                 g = 0;
             }
@@ -381,14 +381,14 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = -CoreRobotics::CR_PI / 2;
+                b = -M_PI / 2;
                 a = -atan2(r10, r11);
                 g = 0;
             }
         } 
         else 
         {
-            b = CoreRobotics::CR_PI / 2;
+            b = M_PI / 2;
             a = atan2(r10, r11);
             g = 0;
         }
@@ -404,14 +404,14 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI / 2;
+                b = M_PI / 2;
                 a = -atan2(r21, r22);
                 g = 0;
             }
         }
         else
         {
-            b = CoreRobotics::CR_PI / 2;
+            b = M_PI / 2;
             a = atan2(r21, r22);
             g = 0;
         }
@@ -427,14 +427,14 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI / 2;
+                b = M_PI / 2;
                 a = -atan2(r02, r00);
                 g = 0;
             }
         }
         else
         {
-            b = CoreRobotics::CR_PI / 2;
+            b = M_PI / 2;
             a = atan2(r02, r00);
             g = 0;
         }
@@ -450,14 +450,14 @@ Eigen::Vector3d CRFrame::getOrientation(CREulerMode i_mode)
             }
             else
             {
-                b = CoreRobotics::CR_PI / 2;
+                b = M_PI / 2;
                 a = -atan2(-r20, r22);
                 g = 0;
             }
         }
         else
         {
-            b = CoreRobotics::CR_PI / 2;
+            b = M_PI / 2;
             a = atan2(-r20, r22);
             g = 0;
[…]
