### 2017-12-19 c899217
removed CR_PI line
