+++ src/math/CRMath.hpp
@@ -93,11 +93,6 @@ John Wiley & Sons, 2011.
 */
 
 //=====================================================================
-
-//! CoreRobotics value for PI
-// const double CR_PI = M_PI;
-
-//=====================================================================
 class CRMath {
 
 //---------------------------------------------------------------------
