### 2017-12-19 774c5a1
fixed refs to CR_PI
