### 2017-12-21 e3bbd35
matlab swig wrappers [ci skip]
