+++ matlab/lib/+CoreRobotics/CRFrameDh.m
@@ -7,41 +7,41 @@ classdef CRFrameDh < CoreRobotics.CRFrame
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(119, varargin{:});
+        tmp = CoreRoboticsMEX(118, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = setFreeValue(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(120, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(119, self, varargin{:});
     end
     function varargout = getFreeValue(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(121, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(120, self, varargin{:});
     end
     function varargout = setFreeVariable(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(122, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(121, self, varargin{:});
     end
     function varargout = getFreeVariable(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(123, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(122, self, varargin{:});
     end
     function varargout = setMode(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(124, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(123, self, varargin{:});
     end
     function varargout = getMode(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(125, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(124, self, varargin{:});
     end
     function varargout = setParameters(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(126, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(125, self, varargin{:});
     end
     function varargout = getParameters(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(127, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(126, self, varargin{:});
     end
     function varargout = isDriven(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(128, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(127, self, varargin{:});
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(129, self);
+        CoreRoboticsMEX(128, self);
         self.swigPtr=[];
       end
     end
