+++ matlab/lib/+CoreRobotics/CRManipulator.m
@@ -9,53 +9,53 @@ classdef CRManipulator < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(135, varargin{:});
+        tmp = CoreRoboticsMEX(134, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = setConfiguration(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(136, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(135, self, varargin{:});
     end
     function varargout = getConfiguration(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(137, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(136, self, varargin{:});
     end
     function varargout = getForwardKinematics(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(138, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(137, self, varargin{:});
     end
     function varargout = getNumberOfLinks(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(139, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(138, self, varargin{:});
     end
     function varargout = getDegreesOfFreedom(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(140, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(139, self, varargin{:});
     end
     function varargout = getToolFrame(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(141, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(140, self, varargin{:});
     end
     function varargout = getLinkFrame(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(142, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(141, self, varargin{:});
     end
     function varargout = getToolPose(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(143, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(142, self, varargin{:});
     end
     function varargout = setModelType(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(144, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(143, self, varargin{:});
     end
     function varargout = jacobian(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(145, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(144, self, varargin{:});
     end
     function varargout = hessian(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(146, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(145, self, varargin{:});
     end
     function varargout = addLink(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(147, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(146, self, varargin{:});
     end
     function varargout = addTool(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(148, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(147, self, varargin{:});
     end
     function delete(self)
[…]

[…]
