### 2018-01-01 #489 @danielskatz
@amjaeger17 - If the checkboxes in your list weren't checked by you, they were accidentally checked by @bmagyar; feel free to uncheck them.

@CoreRobotics - Any news on the issues opened by @bmagyar ?
