+++ src/math/CRMatrix.cpp
@@ -39,7 +39,7 @@ POSSIBILITY OF SUCH DAMAGE.
 */
 //=====================================================================
 
-#include "CRMath.hpp"
+#include "CRMatrix.hpp"
 #include "Eigen/Dense"
 #include "Eigen/SVD"
 
@@ -47,76 +47,6 @@ POSSIBILITY OF SUCH DAMAGE.
 //=====================================================================
 // CoreRobotics namespace
 namespace CoreRobotics {
-
-    
-//=====================================================================
-/*!
- This method performs a forward euler integration step for the dynamical
- system:\n
- 
- \f[
- \dot{x} = f(t,x,u)
- \f]
- 
- This equation is specified as a callback function to the integration.
- 
- \param [in] i_dynamicSystem - the continuous time dynamic system.
- \param [in] i_t - the current time in seconds
- \param [in] i_x - the current state vector
- \param [in] i_u - the current input vector
- \param [in] i_dt - the time step to integrate over
- \return the state vector at t+dt
- 
- */
-//---------------------------------------------------------------------
-Eigen::VectorXd CRMath::forwardEulerStep(Eigen::VectorXd(i_dynamicSystem)(double,
-                                                                          Eigen::VectorXd,
-                                                                          Eigen::VectorXd),
-                                         double i_t,
-                                         Eigen::VectorXd i_x,
-                                         Eigen::VectorXd i_u,
-                                         double i_dt){
-    // forward integration step
-    return i_x + i_dt*i_dynamicSystem(i_t,i_x,i_u);
-}
-    
-    
-//=====================================================================
-/*!
- This method performs a Runge-Kutta integration step for the dynamical
- system:\n
- 
- \f[
- \dot{x} = f(t,x,u)
- \f]
- 
- This equation is specified as a callback function to the integration.
- The Runga-Kutta integration is an accurate integration scheme at the
- expense of computational complexity over the forward Euler method.
- 
- \param [in] i_dynamicSystem - the continuous time dynamic system
- \param [in] i_t - the current time in seconds
- \param [in] i_x - the current state vector
- \param [in] i_u - the current input vector
- \param [in] i_dt - the time step to integrate over
- \return the state vector at t+dt
- 
- */
-//---------------------------------------------------------------------
-Eigen::VectorXd CRMath::rungeKuttaStep(Eigen::VectorXd(i_dynamicSystem)(double,
-                                                                        Eigen::VectorXd,
-                                                                        Eigen::VectorXd),
-                                       double i_t,
-                                       Eigen::VectorXd i_x,
-                                       Eigen::VectorXd i_u,
-                                       double i_dt){
-    // RK4 step
-    Eigen::VectorXd f1 = i_dynamicSystem(i_t,i_x,i_u);
-    Eigen::VectorXd f2 = i_dynamicSystem(i_t+i_dt/2,i_x+i_dt*f1/2,i_u);
-    Eigen::VectorXd f3 = i_dynamicSystem(i_t+i_dt/2,i_x+i_dt*f2/2,i_u);
-    Eigen::VectorXd f4 = i_dynamicSystem(i_t+i_dt,i_x+i_dt*f3,i_u);
-    return i_x + i_dt/6*(f1 + 2*f2 + 2*f3 + f4);
-}
 
 
 
@@ -142,7 +72,7 @@ singular values were below the specified tolerance (i.e. singular)
 
 */
 //---------------------------------------------------------------------
[…]
