### 2018-01-04 CoreRobotics/CoreRobotics#1 @CoreRobotics
Unfortunately, we can't put any figures of our ongoing projects in the paper due to confidentiality issues, but I will update the paper.md with a description of a motivating application.
