### 2018-01-04 16b8072
updated paper with application
