+++ matlab/lib/+CoreRobotics/CRHardLimits.m
@@ -9,80 +9,80 @@ classdef CRHardLimits < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(240, varargin{:});
+        tmp = CoreRoboticsMEX(244, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = getIKSolver(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(241, self, varargin{:});
-    end
-    function varargout = getNullSpaceSolver(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(242, self, varargin{:});
-    end
-    function varargout = useNullSpace(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(243, self, varargin{:});
-    end
-    function varargout = nullSpaceStatus(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(244, self, varargin{:});
-    end
-    function varargout = setPoseElements(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(245, self, varargin{:});
     end
-    function varargout = getPoseElements(self,varargin)
+    function varargout = getNullSpaceSolver(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(246, self, varargin{:});
     end
-    function varargout = setJointUpperLimit(self,varargin)
+    function varargout = useNullSpace(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(247, self, varargin{:});
     end
-    function varargout = setJointLowerLimit(self,varargin)
+    function varargout = nullSpaceStatus(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(248, self, varargin{:});
     end
-    function varargout = getJointUpperLimit(self,varargin)
+    function varargout = setPoseElements(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(249, self, varargin{:});
     end
-    function varargout = getJointLowerLimit(self,varargin)
+    function varargout = getPoseElements(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(250, self, varargin{:});
     end
-    function varargout = setJointLimits(self,varargin)
+    function varargout = setJointUpperLimit(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(251, self, varargin{:});
     end
-    function varargout = setJointUpperLimits(self,varargin)
+    function varargout = setJointLowerLimit(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(252, self, varargin{:});
     end
-    function varargout = setJointLowerLimits(self,varargin)
+    function varargout = getJointUpperLimit(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(253, self, varargin{:});
     end
-    function varargout = getJointUpperLimits(self,varargin)
+    function varargout = getJointLowerLimit(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(254, self, varargin{:});
     end
-    function varargout = getJointLowerLimits(self,varargin)
+    function varargout = setJointLimits(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(255, self, varargin{:});
     end
-    function varargout = setQ0(self,varargin)
+    function varargout = setJointUpperLimits(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(256, self, varargin{:});
     end
-    function varargout = getQ0(self,varargin)
[…]
