+++ matlab/lib/+CoreRobotics/CRRigidBody.m
@@ -9,27 +9,27 @@ classdef CRRigidBody < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(129, varargin{:});
+        tmp = CoreRoboticsMEX(133, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = setFrame(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(130, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(134, self, varargin{:});
     end
     function varargout = m_frame(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(131, self);
+        varargout{1} = CoreRoboticsMEX(135, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(132, self, varargin{1});
+        CoreRoboticsMEX(136, self, varargin{1});
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(133, self);
+        CoreRoboticsMEX(137, self);
         self.swigPtr=[];
       end
     end
