+++ matlab/lib/+CoreRobotics/uniformParam.m
@@ -7,20 +7,20 @@ classdef uniformParam < SwigRef
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(182, self);
+        varargout{1} = CoreRoboticsMEX(186, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(183, self, varargin{1});
+        CoreRoboticsMEX(187, self, varargin{1});
       end
     end
     function varargout = b(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(184, self);
+        varargout{1} = CoreRoboticsMEX(188, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(185, self, varargin{1});
+        CoreRoboticsMEX(189, self, varargin{1});
       end
     end
     function self = uniformParam(varargin)
@@ -29,14 +29,14 @@ classdef uniformParam < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(186, varargin{:});
+        tmp = CoreRoboticsMEX(190, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(187, self);
+        CoreRoboticsMEX(191, self);
         self.swigPtr=[];
       end
     end

[…]
