### 2018-01-04 CoreRobotics/CoreRobotics#1 @CoreRobotics
The paper.md has been updated.
