### 2018-01-04 #477 @danielskatz
@CoreRobotics - this issue is closed - the review is going on in #489
