+++ python/lib/CoreRobotics.py
@@ -1,12 +1,8 @@
 # This file was automatically generated by SWIG (http://www.swig.org).
-# Version 3.0.11
+# Version 4.0.0
 #
 # Do not make changes to this file unless you know what you are doing--modify
 # the SWIG interface file instead.
-
-
-
-
 
 from sys import version_info as _swig_python_version_info
 if _swig_python_version_info >= (2, 7, 0):
@@ -30,17 +26,18 @@ elif _swig_python_version_info >= (2, 6, 0):
         except ImportError:
             import _CoreRobotics
             return _CoreRobotics
-        if fp is not None:
-            try:
-                _mod = imp.load_module('_CoreRobotics', fp, pathname, description)
-            finally:
+        try:
+            _mod = imp.load_module('_CoreRobotics', fp, pathname, description)
+        finally:
+            if fp is not None:
                 fp.close()
-            return _mod
+        return _mod
     _CoreRobotics = swig_import_helper()
     del swig_import_helper
 else:
     import _CoreRobotics
 del _swig_python_version_info
+
 try:
     _swig_property = property
 except NameError:
@@ -108,7 +105,8 @@ class SwigPyIterator(_object):
         raise AttributeError("No constructor defined - class is abstract")
     __repr__ = _swig_repr
     __swig_destroy__ = _CoreRobotics.delete_SwigPyIterator
-    __del__ = lambda self: None
+    def __del__(self):
+        return None
 
     def value(self):
         return _CoreRobotics.SwigPyIterator_value(self)
@@ -271,7 +269,8 @@ class vectorMatrixXd(_object):
     def capacity(self):
         return _CoreRobotics.vectorMatrixXd_capacity(self)
     __swig_destroy__ = _CoreRobotics.delete_vectorMatrixXd
-    __del__ = lambda self: None
+    def __del__(self):
+        return None
 vectorMatrixXd_swigregister = _CoreRobotics.vectorMatrixXd_swigregister
 vectorMatrixXd_swigregister(vectorMatrixXd)
 
@@ -384,7 +383,8 @@ class vectorVectorXd(_object):
     def capacity(self):
         return _CoreRobotics.vectorVectorXd_capacity(self)
     __swig_destroy__ = _CoreRobotics.delete_vectorVectorXd
-    __del__ = lambda self: None
+    def __del__(self):
+        return None
 vectorVectorXd_swigregister = _CoreRobotics.vectorVectorXd_swigregister
 vectorVectorXd_swigregister(vectorVectorXd)
 
@@ -402,7 +402,8 @@ class CRClock(_object):
         except __builtin__.Exception:
             self.this = this
     __swig_destroy__ = _CoreRobotics.delete_CRClock
-    __del__ = lambda self: None
+    def __del__(self):
+        return None
 
     def startTimer(self):
         return _CoreRobotics.CRClock_startTimer(self)
@@ -450,7 +451,8 @@ class CRConversion(_object):
         except __builtin__.Exception:
             self.this = this
     __swig_destroy__ = _CoreRobotics.delete_CRConversion
-    __del__ = lambda self: None
+    def __del__(self):
+        return None
 CRConversion_swigregister = _CoreRobotics.CRConversion_swigregister
 CRConversion_swigregister(CRConversion)
 
@@ -480,7 +482,8 @@ class CRIntegration(_object):
         except __builtin__.Exception:
             self.this = this
[…]
