### 2018-01-04 #477 @CoreRobotics
Ok thanks - I didn't realize I could update the pdf in the new thread.
