### 2018-01-04 #489 @CoreRobotics
@danielskatz We are currently working to address some of the issues @bmagyar opened.  I do have a question... a couple of the issues seem to be preferential, e.g.: #7.  It is difficult at this stage for us to rename all the classes since this would require significant rework of existing projects using the library.  Is this a necessary fix?  Thanks for the clarity.
