### 2018-01-04 CoreRobotics/CoreRobotics#5 @CoreRobotics
This is updated in the joss_issues branch and will be pushed to master shortly.
