### 2018-01-04 CoreRobotics/CoreRobotics#1 @bmagyar
Where? Please reference a pull request. I can't seem to find it on master or in the joss-issues branch.
