### 2018-01-04 CoreRobotics/CoreRobotics#5 @bmagyar
Reference it please in a pull request and explain how it was addressed.

These issues are to keep track of what got done, don't close them unless you've actually merged the solutions.
