### 2018-01-04 b5a0a5d
attempting to switch from caching to artifacts for better reliablility of CI
