+++ .gitlab-ci.yml
@@ -6,33 +6,19 @@ stages:
  - test
  - deploy
 
-python2-wrappers:
+python-wrappers:
   stage: wrap
   script:
    - mkdir build
    - cd build
    - cmake ../ -Dpython=only-swig
    - make
-  cache:
-    key: python2-noswig-$CI_COMMIT_SHA
-    policy: push
+  artifacts:
+    name: python-wrappers-$CI_COMMIT_SHA
     paths:
-     - $CI_PROJECT_DIR/python/swig
-     - $CI_PROJECT_DIR/python/lib
-
-python3-wrappers:
-  stage: wrap
-  script:
-   - mkdir build
-   - cd build
-   - cmake ../ -Dpython=only-swig
-   - make
-  cache:
-    key: python3-noswig-$CI_COMMIT_SHA
-    policy: push
-    paths:
-     - $CI_PROJECT_DIR/python/swig
-     - $CI_PROJECT_DIR/python/lib
+     - $CI_PROJECT_DIR/python/swig/CoreRoboticsPYTHON_wrap.cxx
+     - $CI_PROJECT_DIR/python/lib/CoreRobotics.py
+    expire_in: 1 week
 
 matlab-wrappers:
   stage: wrap
@@ -41,12 +27,13 @@ matlab-wrappers:
    - cd build
    - cmake ../ -Dmatlab=only-swig
    - make
-  cache:
-    key: matlab-$CI_COMMIT_SHA
-    policy: push
+  artifacts:
+    name: matlab-wrappers-$CI_COMMIT_SHA
     paths:
-     - $CI_PROJECT_DIR/matlab/swig
-     - $CI_PROJECT_DIR/matlab/lib
+     - $CI_PROJECT_DIR/matlab/swig/CoreRoboticsMATLAB_wrap.cxx
+     - $CI_PROJECT_DIR/matlab/lib/*.m
+     - $CI_PROJECT_DIR/matlab/lib/+CoreRobotics/*.m
+    expire_in: 1 week
 
 build-cpp:
   stage: build
@@ -55,11 +42,11 @@ build-cpp:
    - cd build
    - cmake ../
    - make
-  cache:
-    key: cpp-$CI_COMMIT_SHA
-    policy: push
+  artifacts:
+    name: cpp-bin-$CI_COMMIT_SHA
     paths:
-     - $CI_PROJECT_DIR/bin
+     - $CI_PROJECT_DIR/bin/TestModules
+    expire_in: 1 week
 
 build-python2:
   stage: build
@@ -68,13 +55,11 @@ build-python2:
    - cd build
    - cmake ../ -Dpython=2
    - make
-  cache:
-    key: python2-$CI_COMMIT_SHA
-    policy: push
+  artifacts:
+    name: python2-bin-$CI_COMMIT_SHA
     paths:
-     - $CI_PROJECT_DIR/python/swig
-     - $CI_PROJECT_DIR/python/lib
-     - $CI_PROJECT_DIR/python/util
+     - $CI_PROJECT_DIR/python/lib/_CoreRobotics.so
+    expire_in: 1 week
 
 build-python3:
   stage: build
@@ -83,13 +68,11 @@ build-python3:
    - cd build
    - cmake ../ -Dpython=3
    - make
-  cache:
-    key: python3-$CI_COMMIT_SHA
-    policy: push
+  artifacts:
+    name: python3-bin-$CI_COMMIT_SHA
     paths:
-     - $CI_PROJECT_DIR/python/swig
-     - $CI_PROJECT_DIR/python/lib
-     - $CI_PROJECT_DIR/python/util
+     - $CI_PROJECT_DIR/python/lib/_CoreRobotics.so
+    expire_in: 1 week
 
[…]
