### 2018-01-04 8e22ef9
added a few paths that were missing from some artifacts, and removed all dependencies from some jobs
