### 2018-01-04 d46751a
removed some redundant dependencies
