### 2018-01-05 cac7f08
Removed dependencies from external folder, and moved to a new repository
