### 2018-01-05 87db8c6
Updated main header for doc updated for new dependencies:
