### 2018-01-05 be6502b
Update README.md - changed eigen3 to boost
