### 2018-01-05 #489 @bmagyar
:+1: for bringing it up here as well, the reviewer guidelines are not specific about this.

While #7 is a pedantic issue, the rest I believe directly affect how others can use your library. Since the point of JOSS is to advocate re-using research code I think it is necessary that the code is brought in line with common design principles.

Of course with open source projects everything is up for discussion.
