### 2018-01-05 32d2681
for wrappers to compile successfully Eigen 3.3 or higher is required. Any version of Eigen is sufficient for compiling the C++ library, although I have had trouble with using apt-get to install eigen3. Boost::Interprocess was introduced in Boost 1.48, so that is the minimum version now required.
