### 2018-01-05 ccb3dfd
removed referances to CR_PI
