+++ python/lib/CoreRobotics.py
@@ -1,8 +1,12 @@
 # This file was automatically generated by SWIG (http://www.swig.org).
-# Version 4.0.0
+# Version 3.0.11
 #
 # Do not make changes to this file unless you know what you are doing--modify
 # the SWIG interface file instead.
+
+
+
+
 
 from sys import version_info as _swig_python_version_info
 if _swig_python_version_info >= (2, 7, 0):
@@ -26,18 +30,17 @@ elif _swig_python_version_info >= (2, 6, 0):
         except ImportError:
             import _CoreRobotics
             return _CoreRobotics
-        try:
-            _mod = imp.load_module('_CoreRobotics', fp, pathname, description)
-        finally:
-            if fp is not None:
+        if fp is not None:
+            try:
+                _mod = imp.load_module('_CoreRobotics', fp, pathname, description)
+            finally:
                 fp.close()
-        return _mod
+            return _mod
     _CoreRobotics = swig_import_helper()
     del swig_import_helper
 else:
     import _CoreRobotics
 del _swig_python_version_info
-
 try:
     _swig_property = property
 except NameError:
@@ -105,8 +108,7 @@ class SwigPyIterator(_object):
         raise AttributeError("No constructor defined - class is abstract")
     __repr__ = _swig_repr
     __swig_destroy__ = _CoreRobotics.delete_SwigPyIterator
-    def __del__(self):
-        return None
+    __del__ = lambda self: None
 
     def value(self):
         return _CoreRobotics.SwigPyIterator_value(self)
@@ -269,8 +271,7 @@ class vectorMatrixXd(_object):
     def capacity(self):
         return _CoreRobotics.vectorMatrixXd_capacity(self)
     __swig_destroy__ = _CoreRobotics.delete_vectorMatrixXd
-    def __del__(self):
-        return None
+    __del__ = lambda self: None
 vectorMatrixXd_swigregister = _CoreRobotics.vectorMatrixXd_swigregister
 vectorMatrixXd_swigregister(vectorMatrixXd)
 
@@ -383,8 +384,7 @@ class vectorVectorXd(_object):
     def capacity(self):
         return _CoreRobotics.vectorVectorXd_capacity(self)
     __swig_destroy__ = _CoreRobotics.delete_vectorVectorXd
-    def __del__(self):
-        return None
+    __del__ = lambda self: None
 vectorVectorXd_swigregister = _CoreRobotics.vectorVectorXd_swigregister
 vectorVectorXd_swigregister(vectorVectorXd)
 
@@ -402,8 +402,7 @@ class CRClock(_object):
         except __builtin__.Exception:
             self.this = this
     __swig_destroy__ = _CoreRobotics.delete_CRClock
-    def __del__(self):
-        return None
+    __del__ = lambda self: None
 
     def startTimer(self):
         return _CoreRobotics.CRClock_startTimer(self)
@@ -451,8 +450,7 @@ class CRConversion(_object):
         except __builtin__.Exception:
             self.this = this
     __swig_destroy__ = _CoreRobotics.delete_CRConversion
-    def __del__(self):
-        return None
+    __del__ = lambda self: None
 CRConversion_swigregister = _CoreRobotics.CRConversion_swigregister
 CRConversion_swigregister(CRConversion)
 
@@ -482,8 +480,7 @@ class CRIntegration(_object):
         except __builtin__.Exception:
             self.this = this
[…]
