### 2018-01-05 d86c998
Eigen version was incorrectly required to be 3.3 not 3.3.3
