### 2018-01-06 8169a0b
changed CI config and combined the two deploy jobs
