### 2018-01-06 8cfea88
moved back to the old configuration but with the last two jobs combined
