### 2018-01-06 f623daa
Added GTest to find_package in CMakeLists.txt
