### 2018-01-07 CoreRobotics/CoreRobotics#5 @CoreRobotics
Alright, since I've already merged the changes back into the master, unfortunately I can't put in a pull request.  I'll reference the commits so you can still see how we addressed the issues.  I'll make sure for the remaining work, I keep it in our joss_issues branch so I can make a pull request and be better at collaboration.

In order to address this issue, we created 3 new files: CRConversion, CRIntegration, and CRMatrix.  Conversion contains static conversion methods (angle conversion, etc...), Integration contains the numerical integration methods, and Matrix contains norms and creation of special matrices (i.e. rotation matrices).  Since these methods are used by other classes, this change affected other files.

See this commit for these changes: https://github.com/CoreRobotics/CoreRobotics/commit/4ffcc17ba431ff83464c797aca049116cd14478f
