### 2018-01-07 CoreRobotics/CoreRobotics#4 @CoreRobotics
Like the others, since I've already merged the changes back into the master, unfortunately I can't put in a pull request. I'll reference the commits so you can still see how we addressed the issues. I'll make sure for the remaining work, I keep it in our joss_issues branch so I can make a pull request and be better at collaboration.

In order to address this issue, we removed all references in the classes and tests to CR_PI.  When a value for pi is needed, we use M_PI from the standard math library constants.  This involved a couple changes to the python and matlab wrappers as well in later commits for CI improvement

See: https://github.com/CoreRobotics/CoreRobotics/commit/058af7bdd931c8f025860ebfb2a3422b6aa349a1 https://github.com/CoreRobotics/CoreRobotics/commit/774c5a1c5fa022535a6df67dc5073e960429822c https://github.com/CoreRobotics/CoreRobotics/commit/ccb3dfd0b2703060623c7f69fd5a8b5cdbbcbcdd
