### 2018-01-07 CoreRobotics/CoreRobotics#3 @CoreRobotics
Like the others, since I've already merged the changes back into the master, unfortunately I can't put in a pull request. I'll reference the commits so you can still see how we addressed the issues. I'll make sure for the remaining work, I keep it in our joss_issues branch so I can make a pull request and be better at collaboration.

Good call on this one.  In order to address this issue, we removed the external folder containing eigen and boost and moved them over to a different repository (https://gitlab.com/powan/CRexternal).  To check for eigen and boost, we updated CMakeLists.txt to automatically find the needed packages.  We also added eigen and boost to the docker image used as our CI runner.  Instructions were adjusted for the new approach.

See: https://github.com/CoreRobotics/CoreRobotics/commit/f7bcef38aa93e6e896c73e08f682baaf60eb941d https://github.com/CoreRobotics/CoreRobotics/commit/32d2681365521b055541a38ee813f8f4696a9baa
