### 2018-01-07 CoreRobotics/CoreRobotics#1 @CoreRobotics
Like the others, since I've already merged the changes back into the master, unfortunately I can't put in a pull request. I'll reference the commits so you can still see how we addressed the issues. I'll make sure for the remaining work, I keep it in our joss_issues branch so I can make a pull request and be better at collaboration.

I've updated the paper.md with a paragraph about an example application.  Although now I'm thinking that it would be good to include a couple example .cpp files that demonstrate how to use our core classes since we are changing the test scripts to use GTest, addressing: https://github.com/CoreRobotics/CoreRobotics/issues/2

See: https://github.com/CoreRobotics/CoreRobotics/commit/16b8072b5419fb2813ccc78e56fea7336ecc1c80
