### 2018-01-07 CoreRobotics/CoreRobotics#2 @CoreRobotics
We're in the process of changing the test scripts to use GTest.  If GTest is not found during the CMake check, then the tests won't build but the rest of the library will.

See: https://github.com/CoreRobotics/CoreRobotics/commit/f623daab14457bb961803dbb0b78843b027a0660
