### 2018-01-07 3547f01
Updated testmodules to use GTest
