### 2018-01-07 e16c14f
Added Core and physics tests
