### 2018-01-07 272aaa5
Added manipulator tests
