### 2018-01-09 aea9b39
modified CRNullSpace test script to use GTest
