+++ utils/test_CRNullSpace.cpp
@@ -41,13 +41,22 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #include <iostream>
 #include "CoreRobotics.hpp"
+#include "gtest/gtest.h"
 
 using namespace CoreRobotics;
 
-void test_CRNullSpace(void) {
+struct testData {
+	Eigen::VectorXd initialToolPose;
+	Eigen::VectorXd finalToolPose;
+	Eigen::VectorXd initialJointAngles;
+	Eigen::VectorXd finalJointAngles;
+	CRResult result; };
 
-	std::cout << "*************************************\n";
-	std::cout << "Demonstration of CRNullSpace.\n";
+testData test_CRNullSpace(Eigen::VectorXd initJoints,
+                          Eigen::VectorXd jointMotion,
+                          bool usePoseElements = false,
+                          Eigen::Matrix<bool, 6, 1> poseElements = (Eigen::Matrix<bool, 6, 1>() << true, true, true, true, true, true).finished()) {
+	testData returnStruct;
 
 	CRManipulator* MyRobot = new CRManipulator();
 
@@ -137,42 +146,59 @@ void test_CRNullSpace(void) {
 	int toolIndex = MyRobot->addTool(linkIndex6, Tool);
 
 	// Initialize the solver
-	CRNullSpace nullSpaceSolver = CRNullSpace(MyRobot, toolIndex, CR_EULER_MODE_XYZ);
+	CRNullSpace nullSpaceSolver = CRNullSpace(MyRobot, toolIndex, CR_EULER_MODE_XYZ);	
 
 	// Set the robot orientation
-	Eigen::VectorXd InitJoints(6);
-	InitJoints << 0.0, 2.0, 0.0, -2, 0.0, M_PI / 8.0;
-	MyRobot->setConfiguration(InitJoints);
+	MyRobot->setConfiguration(initJoints);
+	returnStruct.initialJointAngles = initJoints;
 
 	// Compute the initial tool pose
-	Eigen::VectorXd toolPose;
-	toolPose = MyRobot->getToolPose(toolIndex, CR_EULER_MODE_XYZ);
-	std::cout << "Initial Tool Pose\n" << toolPose.transpose() << std::endl;
+	returnStruct.initialToolPose = MyRobot->getToolPose(toolIndex, CR_EULER_MODE_XYZ);
 
 	// Find a nullspace movement
+	Eigen::VectorXd nullJointMotion(6);
+	if (usePoseElements) {
+		returnStruct.result = nullSpaceSolver.solve(jointMotion, initJoints, poseElements, nullJointMotion);
+	} else {
+		returnStruct.result = nullSpaceSolver.solve(jointMotion, initJoints, nullJointMotion);
+	}
+
+	// Compute the final tool pose
+	MyRobot->setConfiguration(initJoints + nullJointMotion);
+	returnStruct.finalJointAngles = initJoints + nullJointMotion;
+	returnStruct.finalToolPose = MyRobot->getToolPose(toolIndex, CR_EULER_MODE_XYZ);
+
+	return returnStruct;
+}
+
+TEST(CRNullSpace, Full) {
+	Eigen::VectorXd initJoints(6);
+	initJoints << 0.0, 2.0, 0.0, -2, 0.0, M_PI / 8.0;
 	Eigen::VectorXd jointMotion(6);
 	jointMotion << 1, 0, 0, 0, 0, 0;
-	Eigen::VectorXd nullJointMotion(6);
-	CRResult result = nullSpaceSolver.solve(jointMotion, InitJoints, nullJointMotion);
[…]
