### 2018-01-09 CoreRobotics/CoreRobotics#4 @bmagyar
Big thumbs up ! :+1:
