### 2018-01-09 CoreRobotics/CoreRobotics#1 @bmagyar
Good idea on adding examples!
You could also spice up the tests with comments that explain some logic, that also helps people wanting to use the library.
