### 2018-01-09 e453823
Fixed a couple issues preventing compile on windows: cannot require eigen3 and boost when manually defining include dirs, and windows lost M_PI just like windows would
