### 2018-01-09 cf1e47d
modified the CRHardLimits test script to use GTest
