### 2018-01-10 7a8d250
changed all matlab and python test scripts to examples
