### 2018-01-10 5659eb9
updated the CMake file with the changes made to the directory structure
