+++ CMakeLists.txt
@@ -55,7 +55,7 @@ set(CMAKE_BUILD_TYPE Release)
 # Define the project name and other CoreRobotics parameters
 set(PROJECT_NAME "CoreRobotics")
 set(LIB_TARGET_NAME ${PROJECT_NAME})
-set(UTIL_TARGET_NAME "TestModules")
+set(TEST_TARGET_NAME "TestModules")
 
 # project title
 project( ${PROJECT_NAME} )
@@ -168,13 +168,17 @@ if ((NOT python AND NOT matlab AND NOT only-swig) OR all OR cpp)
 
 
 	# =====================================================================
-	# 2. Setup the utility target
+	# 2. Setup the tests target
 	#
 	# Source files
-	file(GLOB UTILITIES
-		utils/*.cpp
-		utils/*.hpp
+	file(GLOB TESTS
+		tests/*.cpp
 	)
+
+	# set the path to the CoreRobotics static library
+    set(COREROBOTICS_LIB
+        ${CMAKE_SOURCE_DIR}/lib/${CMAKE_CFG_INTDIR}/$<TARGET_FILE_NAME:${PROJECT_NAME}>
+    )
 
     if(NOT GTEST_FOUND)
         message(STATUS "GTEST not found, not building the test suite")
@@ -185,26 +189,36 @@ if ((NOT python AND NOT matlab AND NOT only-swig) OR all OR cpp)
         # add the gtest includes
         include_directories(${GTEST_INCLUDE_DIRS})
 
-        # Add an executable for the utilities
-        add_executable( ${UTIL_TARGET_NAME} ${UTILITIES})
+        # Add an executable for the tests
+        add_executable( ${TEST_TARGET_NAME} ${TESTS})
 
         # Set the output directory
-        set_target_properties( ${UTIL_TARGET_NAME} PROPERTIES
-            RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/bin
+        set_target_properties( ${TEST_TARGET_NAME} PROPERTIES
+            RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/tests/bin
         )
 
         # Add the library as a build dependency
-        add_dependencies(${UTIL_TARGET_NAME} ${LIB_TARGET_NAME})
-
-        # Set libraries to link
-        set(UTIL_LINK_FILES
-            ${CMAKE_SOURCE_DIR}/lib/${CMAKE_CFG_INTDIR}/$<TARGET_FILE_NAME:${PROJECT_NAME}>
-        )
+        add_dependencies(${TEST_TARGET_NAME} ${LIB_TARGET_NAME})
 
         # Link the library binary
-        target_link_libraries( ${UTIL_TARGET_NAME} ${UTIL_LINK_FILES} ${LINK_FLAGS} ${GTEST_LIBRARIES})
+        target_link_libraries( ${TEST_TARGET_NAME} ${COREROBOTICS_LIB} ${LINK_FLAGS} ${GTEST_LIBRARIES})
 
     endif()
+
+	# Example scripts
+	file(GLOB EXAMPLES
+		examples/*.cpp
+	)
+
+	# Compile each example
+	foreach(EXAMPLE EXAMPLES)
+		add_executable(${EXAMPLE} ${EXAMPLE})
+		set_target_properties( ${EXAMPLE} PROPERTIES
+            RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/examples/bin
+        )
[…]
