### 2018-01-10 1e92012
fixed a couple issues in the cmake
