### 2018-01-10 33fdcfd
removed CR_PI from come c++ examples
