### 2018-01-10 c441dce
updated the gitlab ci config file for the new structure
