### 2018-01-10 8bb52e2
changed the names of all the documentation examples from test_ to example_
