+++ src/math/CRIntegration.hpp
@@ -73,7 +73,7 @@ This class implements numerical integration methods.
  
 ## Example
 This example shows usage of the math functions.
-\include test_CRMath.cpp
+\include example_CRMath.cpp
  
 ## References
 [1] Kreyszig, E., Advanced Engineering Mathematics, Ed.9,
