+++ src/math/CRConversion.hpp
@@ -74,7 +74,7 @@ This class implements vmath conversions.
  
 ## Example
 This example shows usage of the math functions.
-\include test_CRMath.cpp
+\include example_CRMath.cpp
  
 ## References
 [1] Kreyszig, E., Advanced Engineering Mathematics, Ed.9,
