### 2018-01-10 f90ff85
fixed class in math documentation
