+++ src/math/CRConversion.hpp
@@ -59,7 +59,7 @@ namespace CoreRobotics {
 */
 //---------------------------------------------------------------------
 /*!
-\class CRMath
+\class CRConversion
 \ingroup math
 
 \brief This class implements math conversions.
@@ -68,9 +68,9 @@ namespace CoreRobotics {
 ## Description
 This class implements vmath conversions.
  
-- CRMath::deg2rad converts degress to radians
-- CRMath::rad2deg converts radians to degrees.
-- CRMath::wrapToPi wraps angles to +/- $\pi$ radians.
+- CRConversion::deg2rad converts degress to radians
+- CRConversion::rad2deg converts radians to degrees.
+- CRConversion::wrapToPi wraps angles to +/- \f$\pi\f$ radians.
  
 ## Example
 This example shows usage of the math functions.
