### 2018-01-10 b8c41ab
added necessary artifacts for examples and the changed test bin location
