+++ .gitlab-ci.yml
@@ -42,7 +42,8 @@ build-cpp:
    - make
   artifacts:
     paths:
-     - $CI_PROJECT_DIR/bin/TestModules
+     - $CI_PROJECT_DIR/tests/bin/TestModules
+     - $CI_PROJECT_DIR/examples/bin/*
     expire_in: 1 week
 
 build-python2:
