### 2018-01-10 f03880a
removed dependencies from some CI jobs
