### 2018-01-10 59f593b
removed referances to the word test from the matlab and python examples
