### 2018-01-11 3cb9741
Added CMake config files so that users can just use find_package(CoreRobotics REQUIRED) in their own projects once they've build it in cmake
