### 2018-01-11 0d69daa
Fixed a bug in IK setTolerance()
