### 2018-01-11 891dcbb
Fixed issue that attempted to make config files even if only swig wrappers are being build
