+++ matlab/lib/+CoreRobotics/CRInverseKinematics.m
@@ -17,51 +17,60 @@ classdef CRInverseKinematics < SwigRef
     function varargout = setRobot(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(213, self, varargin{:});
     end
-    function varargout = setToolIndex(self,varargin)
+    function varargout = getRobot(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(214, self, varargin{:});
     end
-    function varargout = setEulerMode(self,varargin)
+    function varargout = setToolIndex(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(215, self, varargin{:});
     end
-    function varargout = getEulerMode(self,varargin)
+    function varargout = getToolIndex(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(216, self, varargin{:});
     end
-    function varargout = setTolerance(self,varargin)
+    function varargout = setEulerMode(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(217, self, varargin{:});
     end
-    function varargout = setMaxIter(self,varargin)
+    function varargout = getEulerMode(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(218, self, varargin{:});
     end
-    function varargout = getMaxIter(self,varargin)
+    function varargout = setTolerance(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(219, self, varargin{:});
     end
-    function varargout = setStepSize(self,varargin)
+    function varargout = getTolerance(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(220, self, varargin{:});
     end
-    function varargout = getStepSize(self,varargin)
+    function varargout = setMaxIter(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(221, self, varargin{:});
     end
-    function varargout = setDampingFactor(self,varargin)
+    function varargout = getMaxIter(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(222, self, varargin{:});
     end
-    function varargout = getDampingFactor(self,varargin)
+    function varargout = setStepSize(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(223, self, varargin{:});
     end
-    function varargout = setSingularThresh(self,varargin)
+    function varargout = getStepSize(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(224, self, varargin{:});
     end
-    function varargout = getSingularThresh(self,varargin)
+    function varargout = setDampingFactor(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(225, self, varargin{:});
     end
-    function varargout = getJacInv(self,varargin)
+    function varargout = getDampingFactor(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(226, self, varargin{:});
     end
-    function varargout = solve(self,varargin)
+    function varargout = setSingularThresh(self,varargin)
       [varargout{1:nargout}] = CoreRoboticsMEX(227, self, varargin{:});
+    end
+    function varargout = getSingularThresh(self,varargin)
+      [varargout{1:nargout}] = CoreRoboticsMEX(228, self, varargin{:});
+    end
+    function varargout = getJacInv(self,varargin)
+      [varargout{1:nargout}] = CoreRoboticsMEX(229, self, varargin{:});
+    end
+    function varargout = solve(self,varargin)
+      [varargout{1:nargout}] = CoreRoboticsMEX(230, self, varargin{:});
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(228, self);
+        CoreRoboticsMEX(231, self);
         self.swigPtr=[];
       end
     end
