### 2018-01-11 65e0f00
Fixed small issue in CMakeLists for package installer
