+++ utils/test_CRInverseKinematics.cpp
@@ -40,305 +40,117 @@ POSSIBILITY OF SUCH DAMAGE.
 //=====================================================================
 #include <iostream>
 #include "CoreRobotics.hpp"
+#include "gtest/gtest.h"
 
 // Use the CoreRobotics namespace
 using namespace CoreRobotics;
 
 
-// -------------------------------------------------------------
-void test_CRInverseKinematics(void) {
-
-	std::cout << "*************************************\n";
-	std::cout << "Demonstration of CRInverseKinematics.\n";
-    std::cout << std::fixed; std::cout.precision(4);
+// Setup the CRManipulator robot and return the toolIndex
+int setup3dof(CRManipulator& MyRobot){
+    CRFrameEuler* F0 = new CRFrameEuler();
+    CRFrameEuler* F1 = new CRFrameEuler();
+    CRFrameEuler* F2 = new CRFrameEuler();
+    CRRigidBody* Link0 = new CRRigidBody();
+    CRRigidBody* Link1 = new CRRigidBody();
+    CRRigidBody* Link2 = new CRRigidBody();
     
-    // Set the Euler convention we will use throughout the example
-    // Although CoreRobotics offers the flexibility to choose a
-    // different convention for each method, in general it is good
-    // to adopt the same convention throughout a problem for
-    // consistency.
-    CREulerMode convention = CR_EULER_MODE_XYZ;
-
-	// ------------------------------------------
-	// Create the robot
-
-	// create several rigid body links
-    CRFrameEuler* F0 = new CRFrameEuler(0, 0, 0, 0, 0, 0,
-                                        convention,
-                                        CR_EULER_FREE_ANG_G);
-	CRFrameEuler* F1 = new CRFrameEuler(1, 0, 0, 0, 0, 0,
-                                        convention,
-                                        CR_EULER_FREE_ANG_G);
-	CRFrameEuler* F2 = new CRFrameEuler(2, 0, 0, 0, 0, 0,
-                                        convention,
-                                        CR_EULER_FREE_ANG_G);
-    CRFrameEuler* F3 = new CRFrameEuler(1, 0, 0, 0, 0, 0,
-                                        convention,
-                                        CR_EULER_FREE_NONE);
-	CRRigidBody* Link0 = new CRRigidBody(F0);
-	CRRigidBody* Link1 = new CRRigidBody(F1);
-	CRRigidBody* Link2 = new CRRigidBody(F2);
-    CRRigidBody* Link3 = new CRRigidBody(F3);
+    // Set info for Link 0 and add to MyRobot
+    F0->setFreeVariable(CR_EULER_FREE_ANG_G);
+    F0->setMode(CR_EULER_MODE_XYZ);
+    F0->setPositionAndOrientation(0, 0, 0.5, 0, 0, 0);
+    Link0->setFrame(F0);
+    MyRobot.addLink(Link0);
     
-    // Create a new robot & add the links
+    // Set info for Link 1 and add to MyRobot
+    F1->setFreeVariable(CR_EULER_FREE_ANG_G);
+    F1->setMode(CR_EULER_MODE_XYZ);
+    F1->setPositionAndOrientation(1, 0, 0, 0, 0, 0);
+    Link1->setFrame(F1);
+    MyRobot.addLink(Link1);
+    
+    // Set info for Link 2 and add to MyRobot
+    F2->setFreeVariable(CR_EULER_FREE_ANG_G);
+    F2->setMode(CR_EULER_MODE_XYZ);
+    F2->setPositionAndOrientation(2, 0, 0, 0, 0, 0);
+    Link2->setFrame(F2);
+    MyRobot.addLink(Link2);
+    
[…]
