### 2018-01-16 8bc7830
added trajectory generator test
