+++ utils/test_CRTrajectoryGenerator.cpp
@@ -40,69 +40,112 @@ POSSIBILITY OF SUCH DAMAGE.
 //=====================================================================
 #include <iostream>
 #include "CoreRobotics.hpp"
+#include "gtest/gtest.h"
 
 // Use the CoreRobotics namespace
 using namespace CoreRobotics;
 
-
-// -------------------------------------------------------------
-void test_CRTrajectoryGenerator(void) {
-
-	std::cout << "*************************************\n";
-	std::cout << "Demonstration of CRTrajectoryGenerator.\n";
-    std::cout << std::fixed; std::cout.precision(4);
-    
-    // ------------------------------------------
-    // Define a trajectory generator
+//
+// Test solve method
+//
+TEST(CRTrajectoryGenerator, Solve){
     CRTrajectoryGenerator trajGen;
-    
-    // intialize a structure for reading the outhput of trajGen
     CRWaypoint wp;
     
-    
-    // ------------------------------------------
-    // Initialize a clock to time the solver
-    CRClock timer = CRClock();
-    
-
-	// ------------------------------------------
-	// Initial and final conditions
-    double tf = 1.2;
-    Eigen::Vector2d x0;
+    // Initial conditions
+    Eigen::Vector2d x0, v0, a0;
     x0 << 0, 0;
-    Eigen::Vector2d v0;
     v0 << 0, -2;
-    Eigen::Vector2d a0;
     a0 << 0, 0;
-    Eigen::Vector2d xf;
+    
+    // Final conditions
+    double tf = 1.2;
+    Eigen::Vector2d xf, vf, af;
     xf << -0.4, 0.5;
-    Eigen::Vector2d vf;
     vf << 1.2, 1;
-    Eigen::Vector2d af;
     af << 0, 0;
     
-    // ------------------------------------------
-    // Compute the trajectory
+    // Check the solve function
+    CRResult result = trajGen.solve(x0, v0, a0, xf, vf, af, tf);
+    EXPECT_EQ(result, CR_RESULT_SUCCESS);
+}
+
+
+
+//
+// Test step methods
+//
+TEST(CRTrajectoryGenerator, Step){
+    CRTrajectoryGenerator trajGen;
+    CRWaypoint wp;
+    
+    // Initial conditions
+    Eigen::Vector2d x0, v0, a0;
+    x0 << 0, 0;
+    v0 << 0, -2;
+    a0 << 0, 0;
+    
+    // Final conditions
+    double tf = 1.2;
+    Eigen::Vector2d xf, vf, af;
+    xf << -0.4, 0.5;
+    vf << 1.2, 1;
+    af << 0, 0;
+    
+    // add a timer for running the internal clock
+    CRClock timer;
+    
+    // Check the internal timer
     trajGen.solve(x0, v0, a0, xf, vf, af, tf);
     timer.startTimer();
-
-    // ------------------------------------------
-    // loop
-    printf("t (s) | Position | Velocity | Acceleration\n");
-    double t = 0;
-    while(t <= 1.2) {
-        
-        wp = trajGen.step(t);
-        
-        // output the time and state
-        printf("%.1f | %+.3f, %+.3f | %+.3f, %+.3f | %+.3f, %+.3f \n", t,
[…]
