### 2018-01-16 d960cc0
Added motion model tests
