### 2018-01-16 9319b34
Changed the propbabilistci test for better sampling test
