### 2018-01-16 ea7f4b5
Added sensor tests
