### 2018-01-17 37685f8
Fixed cmake lists and M_PI issue for windows
