+++ CMakeLists.txt
@@ -76,8 +76,8 @@ project( ${PROJECT_NAME} )
 # if(python OR matlab OR all)
 set(EIGEN_VERSION 3.3.3)
 # endif()
-find_package (Eigen3 ${EIGEN_VERSION} REQUIRED)
-find_package (Boost 1.48 REQUIRED)
+find_package (Eigen3 ${EIGEN_VERSION})
+find_package (Boost 1.48)
 find_package (GTest)
 
 if(NOT Boost_INCLUDE_DIR)
