### 2018-01-18 #489 @danielskatz
Hi - can I get an update from everyone on this review?

@bmagyar, @amjaeger17, @CoreRobotics

Can you each tell me if you are working on this, or waiting for someone else, and if the latter, who and what are you waiting for?

Thanks!
