### 2018-01-18 #489 @bmagyar
From the issues I've raised 3 are still outstanding.
