### 2018-01-18 #489 @CoreRobotics
We're almost finished with the most extensive of the issues - we're wrapping our test scripts with google test. There should be a pull request addressing the 3 outstanding issues in the next day or two.
