### 2018-01-23 b874b21
Added noise model tests
