+++ tests/test_CRNoiseUniform.cpp
@@ -38,18 +38,17 @@ POSSIBILITY OF SUCH DAMAGE.
 
 */
 //=====================================================================
-
-
 #include <iostream>
 #include "CoreRobotics.hpp"
+#include "gtest/gtest.h"
 
 // Use the CoreRobotics namespace
 using namespace CoreRobotics;
 
-void test_CRNoiseUniform(void){
-    
-    std::cout << "*************************************\n";
-    std::cout << "Demonstration of test_CRNoiseUniform.\n";
+//
+// Sample from the distribution and check the histogram
+//
+TEST(CRNoiseUniform, Sample){
     
     // define the uniform properties
     Eigen::VectorXd a(1);
@@ -62,28 +61,37 @@ void test_CRNoiseUniform(void){
     uniformNoise.setParameters(a, b);
     
     // initialize parameters for experiments
-    const int nrolls=10000;  // number of experiments
-    const int nstars=100;     // maximum number of stars to distribute
+    const int n=1000000; // number of experiments
     int p[10]={};
     
     // sample the distribution
-    for (int i=0; i<nrolls; ++i) {
+    for (int i=0; i<n; ++i) {
         Eigen::VectorXd v = uniformNoise.sample();
         if ((v(0)>=0.0)&&(v(0)<10.0)) ++p[int(v(0))];
     }
     
-    // print out the result with stars to indicate density
-    std::cout << std::fixed; std::cout.precision(1);
-    for (int i=0; i<10; ++i) {
-        printf("%2i - %2i | ",i,i+1);
+    // compute the expected distribution for each histogram bin
+    double prob[10]={};
+    for (int k=0; k<10; ++k) {
+        double csum = 0;
+        // integrate 100 step between k and k+1
         Eigen::VectorXd point(1);
-        point << double(i);
-        double prob = uniformNoise.probability(point);
-        printf("%6.4f | ",prob);
-        std::cout << std::string(p[i]*nstars/nrolls,'*') << std::endl;
+        for (int j=0; j<100; ++j){
+            point << double(k) + double(j) / 100;
+            csum += uniformNoise.probability(point);
+        }
+        prob[k] = double(n) * csum / 100.0;
     }
+    
+    // We expect the histogram to be close to the binned pdf (1%)
+    EXPECT_NEAR(prob[0], double(p[0]), 0.01 * double(n));
+    EXPECT_NEAR(prob[1], double(p[1]), 0.01 * double(n));
+    EXPECT_NEAR(prob[2], double(p[2]), 0.01 * double(n));
+    EXPECT_NEAR(prob[3], double(p[3]), 0.01 * double(n));
+    EXPECT_NEAR(prob[4], double(p[4]), 0.01 * double(n));
+    EXPECT_NEAR(prob[5], double(p[5]), 0.01 * double(n));
+    EXPECT_NEAR(prob[6], double(p[6]), 0.01 * double(n));
+    EXPECT_NEAR(prob[7], double(p[7]), 0.01 * double(n));
+    EXPECT_NEAR(prob[8], double(p[8]), 0.01 * double(n));
+    EXPECT_NEAR(prob[9], double(p[9]), 0.01 * double(n));
 }
-
-
-
-
