+++ tests/test_CRNoiseModel.cpp
@@ -38,10 +38,9 @@ POSSIBILITY OF SUCH DAMAGE.
 
 */
 //=====================================================================
-
-
 #include <iostream>
 #include "CoreRobotics.hpp"
+#include "gtest/gtest.h"
 
 // Use the CoreRobotics namespace
 using namespace CoreRobotics;
@@ -60,37 +59,51 @@ double pdensity(Eigen::VectorXd x){
     return 2*x(0);
 }
 
-void test_CRNoiseModel(void){
-    
-    std::cout << "*************************************\n";
-    std::cout << "Demonstration of CRNoiseModel.\n";
+//
+// Sample from the distribution and check the histogram
+//
+TEST(CRNoiseModel, Sample){
     
     // initialize a noise model
     CRNoiseModel genericNoise = CRNoiseModel();
-    genericNoise.setParameters(*icdf,*pdensity);
-    
+    genericNoise.setParameters(*icdf, *pdensity);
     
     // initialize parameters for experiments
-    const int nrolls=10000;  // number of experiments
-    const int nstars=100;    // maximum number of stars to distribute
+    const int n=1000000; // number of experiments
     const int nintervals=10; // number of intervals
-    int p[10]={};
+    int p[nintervals]={};
     
     // sample the distribution
-    for (int i=0; i<nrolls; ++i) {
+    for (int i=0; i<n; ++i) {
         Eigen::VectorXd v = genericNoise.sample();
         ++p[int(nintervals*v(0))];
     }
     
-    // print out the result with stars to indicate density
-    std::cout << std::fixed; std::cout.precision(1);
-    for (int i=0; i<nintervals; ++i) {
-        std::cout << float(i)/nintervals << " - " << float(i+1)/nintervals << ": ";
-        printf("%2i - %2i | ",i,i+1);
+    
+    // compute the expected distribution for each histogram bin
+    double prob[10]={};
+    for (int k=0; k<nintervals; k++){
+        double csum = 0;
+        
+        // integrate 100 steps between k and k+1
         Eigen::VectorXd point(1);
-        point << double(i);
-        double prob = genericNoise.probability(point);
-        printf("%4.1f | ",prob);
-        std::cout << std::string(p[i]*nstars/nrolls,'*') << std::endl;
+        for (int j=0; j<100; ++j){
+            point << double(k) + double(j) / 100.0;
+            csum += genericNoise.probability(point) / 100.0;
+        }
+        // compute the probability
+        prob[k] = csum * (double(n) / 100.0);
     }
+    
+    // We expect the histogram to be close to the binned pdf (1%)
+    EXPECT_NEAR(prob[0], double(p[0]), 0.01 * double(n));
+    EXPECT_NEAR(prob[1], double(p[1]), 0.01 * double(n));
+    EXPECT_NEAR(prob[2], double(p[2]), 0.01 * double(n));
+    EXPECT_NEAR(prob[3], double(p[3]), 0.01 * double(n));
+    EXPECT_NEAR(prob[4], double(p[4]), 0.01 * double(n));
+    EXPECT_NEAR(prob[5], double(p[5]), 0.01 * double(n));
+    EXPECT_NEAR(prob[6], double(p[6]), 0.01 * double(n));
+    EXPECT_NEAR(prob[7], double(p[7]), 0.01 * double(n));
+    EXPECT_NEAR(prob[8], double(p[8]), 0.01 * double(n));
+    EXPECT_NEAR(prob[9], double(p[9]), 0.01 * double(n));
 }

[…]
