### 2018-01-23 0993ba8
Removed dirac from noise mixture example
