### 2018-01-23 930a022
removed dirac test
