### 2018-01-23 980412e
removed Noise Dirac from library
