### 2018-01-23 b975ad1
swig wrappers [ci skip]
