+++ matlab/lib/+CoreRobotics/CRWaypoint.m
@@ -7,50 +7,50 @@ classdef CRWaypoint < SwigRef
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(277, self);
+        varargout{1} = CoreRoboticsMEX(268, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(278, self, varargin{1});
+        CoreRoboticsMEX(269, self, varargin{1});
       end
     end
     function varargout = position(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(279, self);
+        varargout{1} = CoreRoboticsMEX(270, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(280, self, varargin{1});
+        CoreRoboticsMEX(271, self, varargin{1});
       end
     end
     function varargout = velocity(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(281, self);
+        varargout{1} = CoreRoboticsMEX(272, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(282, self, varargin{1});
+        CoreRoboticsMEX(273, self, varargin{1});
       end
     end
     function varargout = acceleration(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(283, self);
+        varargout{1} = CoreRoboticsMEX(274, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(284, self, varargin{1});
+        CoreRoboticsMEX(275, self, varargin{1});
       end
     end
     function varargout = jerk(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(285, self);
+        varargout{1} = CoreRoboticsMEX(276, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(286, self, varargin{1});
+        CoreRoboticsMEX(277, self, varargin{1});
       end
     end
     function self = CRWaypoint(varargin)
@@ -59,14 +59,14 @@ classdef CRWaypoint < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(287, varargin{:});
+        tmp = CoreRoboticsMEX(278, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(288, self);
+        CoreRoboticsMEX(279, self);
         self.swigPtr=[];
       end
     end
