+++ matlab/lib/+CoreRobotics/poseElements.m
@@ -1,3 +1,3 @@
 function varargout = poseElements(varargin)
-  [varargout{1:nargout}] = CoreRoboticsMEX(293, varargin{:});
+  [varargout{1:nargout}] = CoreRoboticsMEX(284, varargin{:});
 end
