+++ matlab/lib/+CoreRobotics/uniformParam.m
@@ -7,20 +7,20 @@ classdef uniformParam < SwigRef
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(186, self);
+        varargout{1} = CoreRoboticsMEX(177, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(187, self, varargin{1});
+        CoreRoboticsMEX(178, self, varargin{1});
       end
     end
     function varargout = b(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(188, self);
+        varargout{1} = CoreRoboticsMEX(179, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(189, self, varargin{1});
+        CoreRoboticsMEX(180, self, varargin{1});
       end
     end
     function self = uniformParam(varargin)
@@ -29,14 +29,14 @@ classdef uniformParam < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(190, varargin{:});
+        tmp = CoreRoboticsMEX(181, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(191, self);
+        CoreRoboticsMEX(182, self);
         self.swigPtr=[];
       end
     end
