### 2018-01-23 2cad768
CRManipulator raw pointer changed to const ref for IK solvers
