+++ tests/test_CRNullSpace.cpp
@@ -146,7 +146,7 @@ testData test_CRNullSpace(Eigen::VectorXd initJoints,
 	int toolIndex = MyRobot->addTool(linkIndex6, Tool);
 
 	// Initialize the solver
-	CRNullSpace nullSpaceSolver = CRNullSpace(MyRobot, toolIndex, CR_EULER_MODE_XYZ);	
+	CRNullSpace nullSpaceSolver = CRNullSpace(*MyRobot, toolIndex, CR_EULER_MODE_XYZ);
 
 	// Set the robot orientation
 	MyRobot->setConfiguration(initJoints);
