### 2018-01-23 CoreRobotics/CoreRobotics#6 @CoreRobotics
Created a pull request at [https://github.com/CoreRobotics/CoreRobotics/pull/8](https://github.com/CoreRobotics/CoreRobotics/pull/8).

We changed the CRManipulator raw pointers to const refs.  See commit [https://github.com/CoreRobotics/CoreRobotics/pull/8/commits/2cad7683c99a73e29e9416e574e97c383f85691c](https://github.com/CoreRobotics/CoreRobotics/pull/8/commits/2cad7683c99a73e29e9416e574e97c383f85691c)
