### 2018-01-23 CoreRobotics/CoreRobotics#2 @CoreRobotics
Created a pull request at [https://github.com/CoreRobotics/CoreRobotics/pull/8](https://github.com/CoreRobotics/CoreRobotics/pull/8).

We added in tests using the gtest framework.  Most of the commits in the pull request are related to this.
