### 2018-01-23 CoreRobotics/CoreRobotics#2 @CoreRobotics
@bmagyar  thank you for suggesting the gtest framework - it is a really nice package.

-Parker
