### 2018-01-23 CoreRobotics/CoreRobotics#1 @CoreRobotics
Created a pull request at [https://github.com/CoreRobotics/CoreRobotics/pull/8](https://github.com/CoreRobotics/CoreRobotics/pull/8).

The original "tests" were more of examples with comments on how to use the classes.  These have been moved to an /examples folder and are individually compiled for brevity. See commit: [https://github.com/CoreRobotics/CoreRobotics/pull/8/commits/3a482c39fa1ea9587051f9734c6378b4372b06f4](https://github.com/CoreRobotics/CoreRobotics/pull/8/commits/3a482c39fa1ea9587051f9734c6378b4372b06f4)
