### 2018-01-23 2a52477
Update .gitlab-ci.yml
