### 2018-01-24 7807ffc
Reset CI yml and CMakeLists for including GTest package
