### 2018-01-24 7255dd9
Update README.md for new install procedure.
