+++ README.md
@@ -20,8 +20,8 @@ This git project hosts code for the CoreRobotics open source robotic control lib
 - *doc/* contains the doxygen Doxyfile for generating html documentation.  Please install [doxygen](http://www.stack.nl/~dimitri/doxygen/) and run the doxyfile to generate.  The main documentation page will be *doc/html/index.html*
 - *python/* contains code for the python wrapper.
 - *src/* contains the library source and headers.  "CoreRobotics.hpp" is the main header for the project.  Classes and functions are organized into sub folders.
-- *utils/* contains code for testing library functionality.  Tests for each class are broken out into test_<classname> and should test functionality of each method.  The test binary gets written to *./bin*  These scripts are also useful for seeing examples of how to use a particular function of the library.
-
+- *tests/* contains code for testing library functionality (Using GTest).  Binaries are compiled to tests/bin.  Tests for each class are broken out into test_<classname> and should test functionality of each method.  The test binary gets written to *./bin*  These scripts are also useful for seeing examples of how to use a particular function of the library.
+- *examples/* constains examples.  Binaries are compiled to examples/bin.
 
 ## External Dependencies
 CoreRobotics makes use of Eigen (vector/matrix math) and Boost libraries.  Make sure you have installed these for your system before proceeding.  Only headers from both libraries are needed, so it is not necessary to compile binaries.  If you are using Mac, you can use homebrew to install.  On Linux, you can use apt-get.
@@ -33,6 +33,8 @@ On Linux using aptitude, the install is
 `brew install boost`
 On Linux using aptitude, the install is
 `sudo apt-get install libboost-all-dev`
+3. [Optional] GTest is used for the test suite, found at [https://github.com/google/googletest](https://github.com/google/googletest).  On Linux using aptitude, the install is
+`sudo apt-get install libgtest0 libgtest-dev`
 
 
 ## Building and compiling
@@ -46,34 +48,15 @@ CoreRobotics relies on CMake to compile cross-platform source to platform-specif
 
 
 ## Using the library
-When built using your CMake-generated project, the CoreRobotics creates a ./bin folder which contains the binary test scripts.  You should make sure this executes.  To use the library in your own project, you need to do the following (example using CMake):
-1. Set the path to your CoreRobotics installation using the CMake `set` command.  
-`set (CR_DIR "path/to/corerobotics/root")`
-2. Add the following directories to the header search paths.  
-`include_directories(
-    ${CR_DIR}/src
-    ${CR_DIR}/src/core
-    ${CR_DIR}/src/math
-    ${CR_DIR}/src/models
-    ${CR_DIR}/src/physics
-    ${CR_DIR}/src/controllers
-)`
-3. Add the external dependencies (boost and eigen) to the header search paths.  If eigen and boost are installed properly (i.e. can be found using find_package() in CMake, then use the following commands,
-`find_package (Eigen3 REQUIRED)
-find_package (Boost 1.65.1 REQUIRED)
-include_directories(
-    ${EIGEN3_INCLUDE_DIR}
-    ${Boost_INCLUDE_DIR}
-)`
-otherwise, you must manually specify the path to eigen and boost
-`include_directories(
-    <path to eigen>
-    <path to boost>
-)`
-4. Include the following folder in your linker directory (note if the library is built in debug, the path must be updated accordingly to reflect).  By default, the compiler builds a static library.
-`link_directories(${CR_DIR}/lib/Release)`
-5. Lastly, you must link the corerobotics library after you add your executable (example is shown for Unix, Windows will be CoreRobotics.lib)
[…]
