### 2018-01-25 246a162
fixed the readme and docs to have the right installation instructions for Eigen and GTest
