### 2018-01-25 CoreRobotics/CoreRobotics#2 @bmagyar
Brilliant! Thumbs up for pushing it through, this will help you keep the library alive for long with ease.
