### 2018-01-25 CoreRobotics/CoreRobotics#9 @bmagyar
[joss-review] Missing DOI entries from bib items
Google Scholar doesn't provide this with their bibtex export but JOSS requires the doi field to be filled for bibliography entries. Please look them up and add them.

Ref: https://github.com/openjournals/joss-reviews/issues/489
