### 2018-01-25 #489 @bmagyar
One last issue from me about the DOI entries, everything else has green lights.
