### 2018-01-25 CoreRobotics/CoreRobotics#9 @CoreRobotics
Quite a few of the references are books that seem to not have DOIs or ISBN-A's that I can find, but they do have regular ISBNs.  What is needed for these cases?  I couldn't find anything on the JOSS site about it.
