### 2018-01-25 CoreRobotics/CoreRobotics#9 @bmagyar
Using the ISBN is a good idea. It serves the same purpose after all.
