### 2018-01-25 CoreRobotics/CoreRobotics#9 @CoreRobotics
Is it ok to just use the ISBN field in the bib entry for these cases or does the DOI field need to have a value?
