### 2018-01-25 ee1f619
added bib updates for DOI/IDBN
