### 2018-01-25 CoreRobotics/CoreRobotics#9 @CoreRobotics
When we include the isbn field in the bib, the joss compiler does not display the number.  Putting it in the doi field would create a broken link.  Any workaround?

See: [https://github.com/openjournals/joss-papers/blob/joss.00489/joss.00489/10.21105.joss.00489.pdf](https://github.com/openjournals/joss-papers/blob/joss.00489/joss.00489/10.21105.joss.00489.pdf)
