### 2018-01-26 CoreRobotics/CoreRobotics#9 @danielskatz
If there is a DOI for a work, then it should be part of the .bib entry.  If there is no DOI, then use whatever identifier exists.  If there is no field for it, you can always use note={ISBN: xxxxx}

The intent of this checklist item was to fix the problem where people put in bib entries with minimal information that doesn't include existing identifiers, which are DOIs for most papers.
