### 2018-01-26 #489 @danielskatz
Thanks @bmagyar - it looks like your side of this is almost done, and the path is clear.
