### 2018-01-26 CoreRobotics/CoreRobotics#9 @CoreRobotics
Okay thanks. All the bib entries have DOI or ISBN. ISBNs are not displayed in the generated pdf.  If that's ok, this should be closed.
