### 2018-01-26 CoreRobotics/CoreRobotics#9 @danielskatz
it's ok, but would probably be better if the ISBNs were added as notes, and did display
