### 2018-01-26 2975346
Update paper.bib for isbn display
