+++ paper.bib
@@ -1,5 +1,6 @@
 @book{kreyszig,
-    isbn = {978-0470458365},
+    note={ISBN: 978-0470458365},
+    isbn={978-0470458365},
     address = {Hoboken, NJ},
     author = {E. Kreyszig},
     edition = {10th},
@@ -9,7 +10,8 @@
 }
 
 @book{craig,
-    isbn = {978-0201543612},
+    note={ISBN: 978-0201543612},
+    isbn={978-0201543612},
     author = {J. Craig},
     title = {Introduction to Robotics: Mechanics and Control},
     edition = {3rd},
@@ -18,6 +20,7 @@
 }
 
 @book{murray,
+    note={ISBN: 978-0849379819},
     isbn = {978-0849379819},
     author = {R. Murray, Z. Li, and S. Sastry},
     edition = {1},
@@ -58,6 +61,7 @@ Automation},
 }
 
 @book{thrun,
+    note={ISBN: 978-0262201629},
     isbn = {978-0262201629},
     address = {Cambridge, MA},
     author = {S. Thrun, W. Burgard, and D. Fox},
@@ -68,7 +72,8 @@ Automation},
 }
 
 @book{crassidis,
-    isbn = {9781439839867},
+    note={ISBN: 978-1439839867},
+    isbn = {978-1439839867},
     address = {Boca Raton, FL},
     author = {J. Crassidis and J. Junkins},
     edition = {2},
