### 2018-01-26 d0f7962
Update paper.bib - changed note field to addendum
