+++ paper.bib
@@ -1,5 +1,5 @@
 @book{kreyszig,
-    note={ISBN: 978-0470458365},
+    addendum={ISBN: 978-0470458365},
     isbn={978-0470458365},
     address = {Hoboken, NJ},
     author = {E. Kreyszig},
@@ -10,7 +10,7 @@
 }
 
 @book{craig,
-    note={ISBN: 978-0201543612},
+    addendum={ISBN: 978-0201543612},
     isbn={978-0201543612},
     author = {J. Craig},
     title = {Introduction to Robotics: Mechanics and Control},
@@ -20,7 +20,7 @@
 }
 
 @book{murray,
-    note={ISBN: 978-0849379819},
+    addendum={ISBN: 978-0849379819},
     isbn = {978-0849379819},
     author = {R. Murray, Z. Li, and S. Sastry},
     edition = {1},
@@ -61,7 +61,7 @@ Automation},
 }
 
 @book{thrun,
-    note={ISBN: 978-0262201629},
+    addendum={ISBN: 978-0262201629},
     isbn = {978-0262201629},
     address = {Cambridge, MA},
     author = {S. Thrun, W. Burgard, and D. Fox},
@@ -72,7 +72,7 @@ Automation},
 }
 
 @book{crassidis,
-    note={ISBN: 978-1439839867},
+    addendum={ISBN: 978-1439839867},
     isbn = {978-1439839867},
     address = {Boca Raton, FL},
     author = {J. Crassidis and J. Junkins},
