### 2018-02-01 3440021
Made noise model parameters public
