### 2018-02-01 53a5746
swig wrappers [ci skip]
