+++ matlab/lib/+CoreRobotics/CRSharedMemory.m
@@ -9,28 +9,28 @@ classdef CRSharedMemory < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(262, varargin{:});
+        tmp = CoreRoboticsMEX(270, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(263, self);
+        CoreRoboticsMEX(271, self);
         self.swigPtr=[];
       end
     end
     function varargout = addSignal(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(264, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(272, self, varargin{:});
     end
     function varargout = removeSignal(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(265, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(273, self, varargin{:});
     end
     function varargout = set(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(266, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(274, self, varargin{:});
     end
     function varargout = get(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(267, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(275, self, varargin{:});
     end
   end
   methods(Static)
