### 2018-02-02 c2104d5
Added methods in CRMatrix for accessing sub vectors and matrices
