### 2018-02-02 0cba0cb
Added GMM/GMR to library
