+++ src/models/CRGmm.hpp
@@ -1,0 +1,165 @@
+//=====================================================================
+/*
+Software License Agreement (BSD-3-Clause License)
+Copyright (c) 2017, CoreRobotics.
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions
+are met:
+
+* Redistributions of source code must retain the above copyright
+notice, this list of conditions and the following disclaimer.
+
+* Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+
+* Neither the name of CoreRobotics nor the names of its contributors
+may be used to endorse or promote products derived from this
+software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+POSSIBILITY OF SUCH DAMAGE.
+
+\project CoreRobotics Project
+\url     www.corerobotics.org
+\author  Parker Owan
+
+*/
+//=====================================================================
+
+#ifndef CRGmm_hpp
+#define CRGmm_hpp
+
+//=====================================================================
+// Includes
+#include "Eigen/Dense"
+#include <random>
+#include <vector>
+#include "CRNoiseMixture.hpp"
+#include "CRNoiseGaussian.hpp"
+
+//=====================================================================
+// CoreRobotics namespace
+namespace CoreRobotics {
+    
+//=====================================================================
+/*!
+ \file CRGmm.hpp
+ \brief Implements a class for Gaussian mixture models.
+ */
+//---------------------------------------------------------------------
+/*!
+ \class CRGmm
+ \ingroup models
+ 
+ \brief Implements a class for Gaussian mixture models.
+ 
+ \details
+ ## Description
+ CRGmm implements methods for sampling and modeling noise as a
+ mixture of Gaussian distributions [2-3].  Each specified Gaussian
+ is also accompanied by a weight, indicating the probability of that
+ distribution being selected for sampling of the noise.
+ 
+ - CRGmm::add adds a noise model to the mixture
+ - CRGmm::sample samples from the noise model
+ - CRGmm::probability evaluates the probability
+ - CRGmm::regression performs Gaussian mixture regression
+ 
+ ## Example
+ This example demonstrates use of the CRGmm class.
+ 
+ \include example_CRGmm.cpp
+ 
+ ## References
[…]
