+++ src/models/CRGmm.cpp
@@ -1,0 +1,256 @@
+//=====================================================================
+/*
+Software License Agreement (BSD-3-Clause License)
+Copyright (c) 2017, CoreRobotics.
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions
+are met:
+
+* Redistributions of source code must retain the above copyright
+notice, this list of conditions and the following disclaimer.
+
+* Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+
+* Neither the name of CoreRobotics nor the names of its contributors
+may be used to endorse or promote products derived from this
+software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+POSSIBILITY OF SUCH DAMAGE.
+
+\project CoreRobotics Project
+\url     www.corerobotics.org
+\author  Parker Owan
+
+*/
+//=====================================================================
+
+#include "Eigen/Dense"
+#include "CRGmm.hpp"
+#include "CRMatrix.hpp"
+#include <chrono>
+#include <iostream>
+
+
+//=====================================================================
+// CoreRobotics namespace
+namespace CoreRobotics {
+    
+    
+//=====================================================================
+/*!
+ The constructor creates a noise model.\n
+ 
+ \param[in] i_seed - seed for the random generator.
+ */
+//---------------------------------------------------------------------
+CRGmm::CRGmm(unsigned i_seed) {
+    
+    this->m_seed = i_seed;
+    this->m_generator.seed(this->m_seed);
+    
+}
+    
+    
+CRGmm::CRGmm() {
+    
+    this->randomSeed();
+    
+}
+    
+    
+//=====================================================================
+/*!
+ This method adds a distribution to the mixture model.
+ 
+ \param[in] i_model - a CRNoiseGaussian distribution.
+ \param[in] i_weight - the corresponding weight of the added distribution.
+ */
+//---------------------------------------------------------------------
+void CRGmm::add(CRNoiseGaussian* i_model, double i_weight)
+{
+    
+    this->m_parameters.models.push_back(i_model);
+    this->m_parameters.weights.push_back(i_weight);
+    
+}
+
+
+//=====================================================================
+/*!
+ This method samples a random number from the mixture model.\n
+ 
+ \return - sampled state.
+ */
[…]
