### 2018-02-02 0093393
fixed issue with virtual function in gmm
