### 2018-02-02 3ac655e
swig wrappers [ci skip]
