+++ matlab/lib/+CoreRobotics/CRParamNoiseMixture.m
@@ -7,20 +7,20 @@ classdef CRParamNoiseMixture < SwigRef
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(194, self);
+        varargout{1} = CoreRoboticsMEX(196, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(195, self, varargin{1});
+        CoreRoboticsMEX(197, self, varargin{1});
       end
     end
     function varargout = weights(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(196, self);
+        varargout{1} = CoreRoboticsMEX(198, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(197, self, varargin{1});
+        CoreRoboticsMEX(199, self, varargin{1});
       end
     end
     function self = CRParamNoiseMixture(varargin)
@@ -29,14 +29,14 @@ classdef CRParamNoiseMixture < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(198, varargin{:});
+        tmp = CoreRoboticsMEX(200, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(199, self);
+        CoreRoboticsMEX(201, self);
         self.swigPtr=[];
       end
     end
