+++ matlab/lib/+CoreRobotics/CRNoiseGaussian.m
@@ -7,33 +7,33 @@ classdef CRNoiseGaussian < CoreRobotics.CRNoiseModel
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(174, varargin{:});
+        tmp = CoreRoboticsMEX(176, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = setParameters(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(175, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(177, self, varargin{:});
     end
     function varargout = sample(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(176, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(178, self, varargin{:});
     end
     function varargout = probability(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(177, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(179, self, varargin{:});
     end
     function varargout = m_parameters(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(178, self);
+        varargout{1} = CoreRoboticsMEX(180, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(179, self, varargin{1});
+        CoreRoboticsMEX(181, self, varargin{1});
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(180, self);
+        CoreRoboticsMEX(182, self);
         self.swigPtr=[];
       end
     end
