+++ matlab/lib/+CoreRobotics/CRNoiseModel.m
@@ -9,33 +9,33 @@ classdef CRNoiseModel < SwigRef
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(159, varargin{:});
+        tmp = CoreRoboticsMEX(161, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = setParameters(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(160, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(162, self, varargin{:});
     end
     function varargout = sample(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(161, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(163, self, varargin{:});
     end
     function varargout = probability(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(162, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(164, self, varargin{:});
     end
     function varargout = m_parameters(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(163, self);
+        varargout{1} = CoreRoboticsMEX(165, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(164, self, varargin{1});
+        CoreRoboticsMEX(166, self, varargin{1});
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(165, self);
+        CoreRoboticsMEX(167, self);
         self.swigPtr=[];
       end
     end
