+++ matlab/lib/+CoreRobotics/CRNoiseMixture.m
@@ -7,33 +7,33 @@ classdef CRNoiseMixture < CoreRobotics.CRNoiseModel
           self.swigPtr = varargin{1}.swigPtr;
         end
       else
-        tmp = CoreRoboticsMEX(200, varargin{:});
+        tmp = CoreRoboticsMEX(202, varargin{:});
         self.swigPtr = tmp.swigPtr;
         tmp.swigPtr = [];
       end
     end
     function varargout = add(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(201, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(203, self, varargin{:});
     end
     function varargout = sample(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(202, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(204, self, varargin{:});
     end
     function varargout = probability(self,varargin)
-      [varargout{1:nargout}] = CoreRoboticsMEX(203, self, varargin{:});
+      [varargout{1:nargout}] = CoreRoboticsMEX(205, self, varargin{:});
     end
     function varargout = m_parameters(self, varargin)
       narginchk(1, 2)
       if nargin==1
         nargoutchk(0, 1)
-        varargout{1} = CoreRoboticsMEX(204, self);
+        varargout{1} = CoreRoboticsMEX(206, self);
       else
         nargoutchk(0, 0)
-        CoreRoboticsMEX(205, self, varargin{1});
+        CoreRoboticsMEX(207, self, varargin{1});
       end
     end
     function delete(self)
       if self.swigPtr
-        CoreRoboticsMEX(206, self);
+        CoreRoboticsMEX(208, self);
         self.swigPtr=[];
       end
     end

[…]
