### 2018-02-03 0a9244b
Fixed NULL conversion issue with CRFrame
