### 2018-02-03 #489 @danielskatz
Since @amjaeger17 seems to be missing in action, we may proceed without him.  @amjaeger17 , last chance to speak up...

But I have a few small issues to bring up before we accept this, basically aimed at the fact that repo seems well-written for developers, but not fully for users.  I think the README needs:

- some of the paper content, such as a brief statement of what problems the software is designed to solve and who the target audience is would fit well in the description section
- in the building and compiling section, something that allows the user to know if the build is successful, such as an example that can be run, and what the results should be
- a section called support, that tells users how to report issues or problems with the software, or how else to seek support
