### 2018-02-05 dcdae67
Updated readme with JOSS editor comments
