### 2018-02-05 f05ba92
Fixed M_PI issue on windows
