### 2018-02-05 #489 @danielskatz
Thanks @CoreRobotics - this looks good.

And @bmagyar - thanks for your reviewing work.

@arfon, please move ahead with accepting this
