### 2018-02-05 #489 @arfon
> @arfon OK thanks. Is it ok for us to roll to v1.0 prior the Zenodo release DOI? The version reviewed was 0.9.1 but the changes we've made from this review have prepared for release.

Yeah, I think that's fine.
