### 2018-02-05 29319a6
Rolled version to 1.0.0
