+++ README.md
@@ -7,7 +7,7 @@
 #### College of Engineering
 
 Created: 2016   
-Version: 0.9.1
+Version: 1.0.0
 
 - C++ Platforms: Windows 8.1/10, Mac OS X 10, and Linux.
 
