### 2018-02-05 0cad338
Added realmin offset to avoid zero divide
