### 2018-02-06 #489 @CoreRobotics
@arfon we have updated v1.0.0 release at Zenodo [https://zenodo.org/record/1166214#.Wnj8IZM-dE4](https://zenodo.org/record/1166214#.Wnj8IZM-dE4)
